+++ src/index.ts
@@ -4,5 +4,6 @@
 import PineTS from './PineTS.class';
 import { Context } from './Context.class';
 import { Provider } from './marketData/Provider.class';
+import { Indicator } from './Indicator';
 
-export { PineTS, Context, Provider };
+export { PineTS, Context, Provider, Indicator };